pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "active-info"
version = "0.1.0"
description = "Active information aggregation for AI/IT opportunities"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "apscheduler>=3.10.4",
  "fastapi>=0.115.0",
  "feedparser>=6.0.11",
  "jinja2>=3.1.4",
  "markdown>=3.7",
  "openai>=1.51.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.2",
  "python-dateutil>=2.9.0",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "typer>=0.12.5",
  "uvicorn>=0.31.0",
]

[project.scripts]
active-info = "active_info.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
